HACK: Mark BASE/OFFSET/SIZE/PERMS/MEMCAP as static link-time constants
Currently the runtime linker adds relocbase to all capabilities, so an
extra REL32 should not be emitted. However, this should really be done
more like everything else, which would have a REL32 for BASE as well as
the MCT entries, but that needs more work, since which symbol is which
entry in the MCT is not known until they have all been added due to the
OFF11/OFF32 split. Perhaps that should be abandoned...