Mark __cap_relocs as SHF_WRITE
Could be a bit smarter here and only do it when there are no dynamic
relocations. This ensures that __cap_relocs ends up in RELRO instead of
rodata which was the case after the latest merge.
For static position-dependent binaries we could put it in the read-only
section but this can be done later.