[lld] Support TLS in RISC-V
Summary:
This patch adds support for thread-local storage (TLS). All four TLS models: global/local-dynamic (both are the same in RISC-V), initial-exec and local-exec are supported, but TLS relaxation is not implemented yet.

In RISC-V (in glibc's implementation at least), the thread pointer (tp/x4) points to the end of TCB and the beginning of the static TLS block, and each DTV points to 0x800 past the start of the TLS block.

Reviewers: ruiu, asb, espindola

Reviewed By: ruiu

Subscribers: rprichard, jocewei, jrtc27, shiva0217, MaskRay, zzheng, edward-jones, mgrang, rogfer01, MartinMosbeck, brucehoult, the_o, rkruppe, niosHD, kito-cheng, simoncook, jordy.potman.lists, apazos, sabuasal, arichardson, johnrusso, emaste, rbar

Tags: #lld

Differential Revision: https://reviews.llvm.org/D39324